requirements.txt
torch==https://download.pytorch.org/whl/cpu/torch-1.5.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
numpy==1.17.4
Flask==1.1.2
torchvision==https://download.pytorch.org/whl/cpu/torchvision-0.6.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
Pillow==7.1.2
gunicorn==20.0.4
